Move ASN to a separate module
