+++ netbox/ipam/models/__init__.py
@@ -1,4 +1,5 @@
 # Ensure that VRFs are imported before IPs/prefixes so dumpdata & loaddata work correctly
+from .asns import *
 from .fhrp import *
 from .vrfs import *
 from .ip import *
